Wed, Feb 9, 2022, 2:21 AM +00:00

+++ CS50x/week4/pset4/filter-less/helpers.c
@@ -59,7 +59,7 @@ void sepia(int height, int width, RGBTRIPLE image[height][width])
 // Reflect image horizontally
 void reflect(int height, int width, RGBTRIPLE image[height][width])
 {
-    int buffer[width];
+    BYTE buffer[width];
 
     for (int i = 0; i < height; i++)
     {
